third_party/blink/web_tests/svg/dynamic-updates/SVGFESpecularLightingElement-dom-in-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'in' attribute of the SVGFESpecularLightingElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS specularElement.getAttribute('in') is "SourceAlpha"
PASS specularElement.getAttribute('in') is "SourceGraphic"
PASS successfullyParsed is true

TEST COMPLETE
